pyproject.toml
[project]
name = "acm-summer-school-kare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocoap>=0.4.14",
    "paho-mqtt>=2.1.0",
]
